refactor: move the type file related to faq next to FAQ component
